+++ server/lib/hubspot-client.js
@@ -17,7 +17,7 @@ export default class HubspotClient {
     return req
       .use(prefixPlugin(process.env.OVERRIDE_HUBSPOT_URL || "https://api.hubapi.com"))
       .use(superagentPromisePlugin)
-      .query({ access_token: accessToken })
+      .set("Authorization", `Bearer ${accessToken}`)
       .on("request", (reqData) => {
         this.metric.increment("ship.service_api.call", 1);
         this.client.logger.debug("hubspotClient.req", reqData.url);
@@ -45,11 +45,13 @@ export default class HubspotClient {
       return Promise.reject(new Error("Refresh token is not set."));
     }
     this.metric.increment("ship.service_api.call", 1);
-    return this.attach(this.req.post("/auth/v1/refresh"))
+    return this.attach(this.req.post("/oauth/v1/token"))
       .set("Content-Type", "application/x-www-form-urlencoded")
       .send({
         refresh_token: refreshToken,
         client_id: process.env.CLIENT_ID,
+        client_secret: process.env.CLIENT_SECRET,
+        redirect_uri: "",
         grant_type: "refresh_token"
       });
   }